alert-service/requirements.txt
kafka-python==2.0.2
redis==5.0.7
pymongo==4.9.2
python-dateutil==2.9.0.post0
